requirements.txt
tornado==3.1.1
wsgiref==0.1.2
tstore==0.3
coverage
pysvcmetrics==0.1
pysvclog=0.1
